+++ R/workability.R
@@ -2,30 +2,35 @@
 #'
 #' This function calculates the workability of soils, given as potential length of the growing season. Based on Huinink (2018)
 #' 
-#' @param A_CLAY_MI (numeric) The clay content of the soil (in procent)
-#' @param A_SILT_MI (numeric) The silt content of the soil (in procent)
-#' @param B_LU_BRP (numeric) The crop code (gewascode) from the BRP
-#' @param B_SOILTYPE_AGR (character) The type of soil
+#' @param A_CLAY_MI (numeric) The clay content of the soil (\%)
+#' @param A_SILT_MI (numeric) The silt content of the soil (\%)
+#' @param B_LU_BRP (numeric) The crop code from the BRP
+#' @param B_SOILTYPE_AGR (character) The agricultural type of soil
 #' @param B_GWL_GLG (numeric) The lowest groundwater level averaged over the most dry periods in 8 years in cm below ground level
 #' @param B_GWL_GHG (numeric) The highest groundwater level averaged over the most wet periods in 8 years in cm below ground level
 #' @param B_Z_TWO  (numeric) The distance between ground level and groundwater level at which the groundwater can supply the soil surface with 2mm water per day (in cm)
+#' @param calcyieldloss (boolean) whether the function includes yield loss, options: TRUE or FALSE (default).
 #' 
 #' @import data.table
 #' 
 #' @references Huinink (2018) Bodem/perceel geschiktheidsbeoordeling voor Landbouw, Bosbouw en Recreatie. BodemConsult-Arnhem
 #'  
 #' @export
-calc_workability <- function(A_CLAY_MI, A_SILT_MI, B_LU_BRP, B_SOILTYPE_AGR, B_GWL_GLG, B_GWL_GHG, B_Z_TWO) {
+calc_workability <- function(A_CLAY_MI, A_SILT_MI, B_LU_BRP, B_SOILTYPE_AGR, 
+                             B_GWL_GLG, B_GWL_GHG, B_Z_TWO,
+                             calcyieldloss = FALSE) {
   
   # define variables used within the function
   id =crop_code = soiltype = landuse = crop_waterstress = crop_season = NULL
   soiltype.m = spring_depth =  gws_sub_workingdepth = NULL
   early_season_day_deficit = late_season_day_deficit = NULL
   req_days_pre_glg = req_days_post_glg = total_days = NULL
+  derving = yield = yl = NULL
+  
   # parameters related to required depth (rd): required depth for capilairy (rdc), for hydrostatic equilibrium (rdh)
   rdh = rdc = rd_spring = rd_fall = NULL
   rdh_spring = rdh_fall = req_spring_depth_day = req_fall_depth_day = NULL
-  rsl = derving = yl = yield = NULL
+  rsl = score = NULL
   
   # Load in the datasets
   crops.obic <- as.data.table(OBIC::crops.obic)
@@ -113,8 +118,8 @@ calc_workability <- function(A_CLAY_MI, A_SILT_MI, B_LU_BRP, B_SOILTYPE_AGR, B_G
      req_fall_depth_day := round(138-(asin((-rd_fall-0.5*(-B_GWL_GHG-B_GWL_GLG))/(0.5*(-B_GWL_GHG+B_GWL_GLG)))/0.0172024))]
   
   # if highest groundwater level is deeper than required depth, soil is always workable
-  # if lowest groundwater level is higher than required depth, soil is never workable.
-  # required_depth_day is set to 228 (which is 15 aug/GLG) so season length will be 0
+  # if lowest groundwater level is higher than required depth, soil is never workable
+  # then required_depth_day is set to 228 (which is 15 aug/GLG) so season length will be 0
   dt[rd_spring < B_GWL_GHG, req_spring_depth_day := 1] 
   dt[rd_spring >= B_GWL_GLG, req_spring_depth_day := 228] 
   dt[rd_fall < B_GWL_GHG, req_fall_depth_day := 1] 
@@ -127,33 +132,52 @@ calc_workability <- function(A_CLAY_MI, A_SILT_MI, B_LU_BRP, B_SOILTYPE_AGR, B_G
   # Calculate relative season length
   dt[,rsl := (total_days-late_season_day_deficit-early_season_day_deficit)/total_days]
   
-  # # Calculate percentage yield loss non-grassland by sub-optimal season length
-  dt[derving == 'zaai groenten' , yl := 538*rsl^2-1144*rsl+606]
-  dt[derving == 'zomergranen'   , yl := 232*rsl^2- 475*rsl+243]
-  dt[derving == 'plant groenten', yl := 392*rsl^2- 785*rsl+393]
-  dt[derving == 'wintergranen'  , yl :=(232*rsl^2- 475*rsl+243)*0.85/2]
-  dt[derving == 'boomteelt'     , yl :=(538*rsl^2-1144*rsl+606)/2]
-  dt[derving == 'overig'        , yl := 100*rsl^2- 100*rsl+100]
-  
-  # helper functions to determine yield loss in grass given soiltype
-  ylveen <- approxfun(x = c(1, 0.9, 0.8, 0.6, 0.55, 0.43, 0.22, 0.08, 0),
-                      y = c(23, 25, 28, 36, 38, 41, 54, 59, 100), method = 'linear',
-                      yleft = NA_integer_, yright = NA_integer_)
-  ylsoil <- approxfun(x = c(1, 0.9, 0.8, 0.6, 0.55, 0.43, 0.22, 0.08, 0),
-                      y = c(23, 25, 29, 41, 43, 51, 68, 72, 100), method = 'linear',
-                      yleft = NA_integer_, yright = NA_integer_)
-  
-  dt[derving == 'grasland' & soiltype.m == 'veen', yl := ylveen(rsl)]
-  dt[derving == 'grasland' & !soiltype.m == 'veen', yl := ylsoil(rsl)]
-  
-  # Calculate yield fraction, always above zero
-  dt[,yield := pmax(0, 1 - 0.01 * yl)] 
+  # Calculate percentage yield loss non-grassland by sub-optimal season length
+  if(calcyieldloss == TRUE){
+    
+    # add yield loss per category
+    dt[derving == 'zaai groenten' , yl := 538 * rsl^2-1144 * rsl + 606]
+    dt[derving == 'zomergranen'   , yl := 232 * rsl^2- 475 * rsl + 243]
+    dt[derving == 'plant groenten', yl := 392 * rsl^2- 785 * rsl + 393]
+    dt[derving == 'wintergranen'  , yl :=(232 * rsl^2- 475 * rsl + 243)*0.85/2]
+    dt[derving == 'boomteelt'     , yl :=(538 * rsl^2-1144 * rsl + 606)/2]
+    dt[derving == 'overig'        , yl := 100 * rsl^2- 200 * rsl + 100]
+    
+    # # helper functions to determine yield loss in grass given soil type
+    ylveen <- approxfun(x = c(1, 0.9, 0.8, 0.6, 0.55, 0.43, 0.22, 0.08, 0),
+                        y = c(23, 25, 28, 36, 38, 41, 54, 59, 100), method = 'linear',
+                        yleft = NA_integer_, yright = NA_integer_)
+    ylsoil <- approxfun(x = c(1, 0.9, 0.8, 0.6, 0.55, 0.43, 0.22, 0.08, 0),
+                        y = c(23, 25, 29, 41, 43, 51, 68, 72, 100), method = 'linear',
+                        yleft = NA_integer_, yright = NA_integer_)
+ 
+    # add yield reduction for grassland
+    dt[derving == 'grasland' & soiltype.m == 'veen', yl := ylveen(rsl)]
+    dt[derving == 'grasland' & !soiltype.m == 'veen', yl := ylsoil(rsl)]
+
+    # Calculate yield fraction, always above zero
+    dt[,yield := pmax(0, 1 - 0.01 * yl)] 
+    dt[derving == 'grasland', yield := evaluate_logistic(x = yield, b = 16, x0 = 0.5, 
+                                                         v = 0.5,increasing = TRUE)]
+    
+  }
+    
   
   # setorder
   setorder(dt, id)
   
-  # Return yield
-  value <- dt[,yield]
+  # Return output
+  if(calcyieldloss == TRUE){
+    
+    # return yield decline
+    value <- dt[,yield]
+    
+  } else {
+    
+    # return relative length season
+    value <- dt[,rsl]  
+  }
+  
   
   # return value
   return(value)
@@ -164,14 +188,47 @@ calc_workability <- function(A_CLAY_MI, A_SILT_MI, B_LU_BRP, B_SOILTYPE_AGR, B_G
 #' This function calculates the indicator for the workability of the soil expressed as the period in which the soil can be worked without
 #' inflicting structural damage that cannot be restored by the regular management on the farm.
 #'  
-#' @param D_P_WO (numeric) The value of workability calculated by \code{\link{calc_workability}}
+#' @param D_WO (numeric) The value of the relative (workable) season length calculated by \code{\link{calc_workability}}
+#' @param B_LU_BRP (numeric) The crop code from the BRP
 #'  
 #' @export
-ind_workability <- function(D_P_WO) {
+ind_workability <- function(D_WO, B_LU_BRP) {
+  
+  # add visual bindings
+  id = arg.length = crop_code = crop_season = rsl = . = NULL
+  
+  # Load in the datasets
+  crops.obic <- as.data.table(OBIC::crops.obic)
+  setkey(crops.obic, crop_code)
+  
+  # length of inputs
+  arg.length <- max(length(D_WO), length(B_LU_BRP))
   
   # Check inputs
-  checkmate::assert_numeric(D_P_WO, lower = 0, upper = 1, any.missing = FALSE)
+  checkmate::assert_numeric(D_WO, lower = 0, upper = 1, any.missing = FALSE)
+  checkmate::assert_numeric(B_LU_BRP, any.missing = FALSE, min.len = 1, len = arg.length)
+  checkmate::assert_subset(B_LU_BRP, choices = unique(crops.obic$crop_code), empty.ok = FALSE)
   
-  # Maybe insert logistic function
-  return(D_P_WO)
+  # Form data table
+  dt <- data.table(id = 1:arg.length,
+                   rsl = D_WO,
+                   B_LU_BRP = B_LU_BRP)
+  
+  # Merge crop_season into data table
+  dt <- merge.data.table(dt, crops.obic[,.(crop_code, crop_season)], by.x = 'B_LU_BRP', by.y = 'crop_code')
+  
+  # evaluate relative season length
+  dt[!grepl('^beweid', crop_season),score := evaluate_logistic(x = rsl, b = 15, x0 = 0.75, v = 1, increasing = TRUE)]
+  dt[grepl('^beweid', crop_season),score := evaluate_logistic(x = rsl, b = 9, x0 = 0.5, v = 1, increasing = TRUE)]
+  
+  # overwrite score when rsl = 1
+  dt[rsl == 1, score := 1]
+  
+  # setorder
+  setorder(dt, id)
+  
+  # Return score
+  score <- dt[,score]
+  
+  return(score)
 }